requirements.txt
aioredis==1.2.0
asgiref==3.2.0
astroid==2.2.4
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.2.1
Automat==0.7.0
cairocffi==1.0.2
CairoSVG==2.4.1
certifi==2018.11.29
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
cssselect2==0.2.1
daphne==2.3.0
defusedxml==0.5.0
Django==2.1.7
django-cors-headers==3.0.2
django-filter==2.1.0
django-mysql==3.0.0.post1
django-redis==4.10.0
djangorestframework==3.9.4
GDAL==3.0.0
hiredis==1.0.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
isort==4.3.11
lazy-object-proxy==1.3.1
Markdown==3.1.1
mccabe==0.6.1
msgpack==0.6.1
mysqlclient==1.4.2.post1
numpy==1.16.2
oauth==1.0.1
oauthlib==3.0.1
pdfkit==0.6.1
Pillow==5.4.1
pyBarcode==0.7
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.3.1
PyPDF2==1.26.0
Pyphen==0.9.5
pypiwin32==223
python3-openid==3.1.0
pytz==2018.9
pywin32==224
redis==3.2.1
reportlab==3.5.23
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social==0.1.dev3
social-auth-app-django==3.1.0
social-auth-core==3.1.0
tinycss2==1.0.2
Twisted==18.9.0
txaio==18.8.1
typed-ast==1.3.1
urllib3==1.24.1
virtualenv==16.6.1
webencodings==0.5.1
wrapt==1.11.1
xhtml2pdf==0.2.3
zope.interface==4.6.0